Add island test and minor nit fixes
